Add helpers to utils.go, correct error code responses
