+++ bazel/execution/bazelapi/definitions.go
@@ -5,43 +5,54 @@
 package bazelapi
 
 import (
+	"fmt"
+
 	remoteexecution "google.golang.org/genproto/googleapis/devtools/remoteexecution/v1test"
 
 	bazelthrift "github.com/twitter/scoot/bazel/execution/bazelapi/gen-go/bazel"
 	scootproto "github.com/twitter/scoot/common/proto"
-	"github.com/twitter/scoot/scootapi/gen-go/scoot"
 )
 
 // These types give us single reference points for passing Execute Requests and Action Results,
 // and leaves room for internal modifications if required
 type ExecuteRequest struct {
-	Request remoteexecution.ExecuteRequest
+	Request      remoteexecution.ExecuteRequest
+	ActionDigest remoteexecution.Digest
 }
 
 type ActionResult struct {
-	Result remoteexecution.ActionResult
+	Result       remoteexecution.ActionResult
+	ActionDigest remoteexecution.Digest
 }
 
 func (e *ExecuteRequest) String() string {
 	if e == nil {
 		return ""
 	}
-	return e.Request.String()
+	return fmt.Sprintf("%s\n%s", e.Request, e.ActionDigest)
 }
 
 func (a *ActionResult) String() string {
 	if a == nil {
 		return ""
 	}
-	return a.Result.String()
+	return fmt.Sprintf("%s\n%s", a.Result, a.ActionDigest)
 }
 
 func (a *ActionResult) GetResult() *remoteexecution.ActionResult {
 	if a == nil {
 		return &remoteexecution.ActionResult{}
 	}
-	res := a.Result
-	return &res
+	r := a.Result
+	return &r
+}
+
+func (a *ActionResult) GetActionDigest() *remoteexecution.Digest {
+	if a == nil {
+		return &remoteexecution.Digest{}
+	}
+	ad := a.ActionDigest
+	return &ad
 }
 
 // Transform request Bazel ExecuteRequest data into a domain object
@@ -54,7 +65,11 @@ func MakeExecReqDomainFromThrift(thriftRequest *bazelthrift.ExecuteRequest) *Exe
 		SkipCacheLookup: thriftRequest.GetSkipCache(),
 		Action:          makeActionFromThrift(thriftRequest.GetAction()),
 	}
-	return &ExecuteRequest{Request: er}
+	d := remoteexecution.Digest{} // TODO fill in hash/sizebytes
+	return &ExecuteRequest{
+		Request:      er,
+		ActionDigest: d,
+	}
 }
 
 // Transform domain ExecuteRequest object into request representation
@@ -83,7 +98,11 @@ func MakeActionResultDomainFromThrift(thriftResult *bazelthrift.ActionResult_) *
 		OutputDirectories: makeOutputDirsFromThrift(thriftResult.GetOutputDirectories()),
 		ExitCode:          thriftResult.GetExitCode(),
 	}
-	return &ActionResult{Result: ar}
+	d := remoteexecution.Digest{} // TODO fill in hash/sizebytes
+	return &ActionResult{
+		Result:       ar,
+		ActionDigest: d,
+	}
 }
 
 // Transform domain ActionResult object into thrift representation
@@ -248,30 +267,4 @@ func makePropertiesThriftFromDomain(platform *remoteexecution.Platform) []*bazel
 		props = append(props, bp)
 	}
 	return props
-}
-
-func RunStatusToExecuteOperationMetadata_Stage(rs *scoot.RunStatus) remoteexecution.ExecuteOperationMetadata_Stage {
-	if rs == nil {
-		return remoteexecution.ExecuteOperationMetadata_UNKNOWN
-	}
-	switch rs.Status {
-	case scoot.RunStatusState_UNKNOWN:
-		return remoteexecution.ExecuteOperationMetadata_UNKNOWN
-	case scoot.RunStatusState_PENDING:
-		return remoteexecution.ExecuteOperationMetadata_QUEUED
-	case scoot.RunStatusState_RUNNING:
-		return remoteexecution.ExecuteOperationMetadata_EXECUTING
-	case scoot.RunStatusState_COMPLETE:
-		return remoteexecution.ExecuteOperationMetadata_COMPLETED
-	case scoot.RunStatusState_FAILED:
-		return remoteexecution.ExecuteOperationMetadata_COMPLETED
-	case scoot.RunStatusState_ABORTED:
-		return remoteexecution.ExecuteOperationMetadata_COMPLETED
-	case scoot.RunStatusState_TIMEDOUT:
-		return remoteexecution.ExecuteOperationMetadata_COMPLETED
-	case scoot.RunStatusState_BADREQUEST:
-		return remoteexecution.ExecuteOperationMetadata_COMPLETED
-	default:
-		return remoteexecution.ExecuteOperationMetadata_UNKNOWN
-	}
 }
